Camada Controls: Adicionado a classe RegisterProvidersMethods, para que a mesma receba os métodos de seu respectivo form.
